requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
click==8.1.7
fastapi==0.112.0
h11==0.14.0
idna==3.7
mysql-connector-python==9.0.0
passlib==1.7.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.6
